Se corrigen errores

+++ AppSapiens/src/main/java/com/mdval/bussiness/entities/ValidaParticula.java
@@ -1,5 +1,7 @@
 package com.mdval.bussiness.entities;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Builder(toBuilder = true)
 public class ValidaParticula {
 
-    private Integer numeroParticula;
+    private BigDecimal numeroParticula;
     private String txtValidacion;
     private String txtValor;
     private String descripcionEstadoValidacion;
